move editableTagProperty to tagProperty
